==> 2_Explorando_os_dados.ipynb <==
from pyspark.sql.session import SparkSession, SparkConf
import pyspark.sql.functions as f
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns
plt.rcParams["figure.figsize"] = 10,5


config = SparkConf().setAll([('spark.executor.memory', '4g'), ('spark.executor.cores', '2'), ('spark.cores.max', '2'), ('spark.driver.memory','4g')])

spark = SparkSession.builder.config(conf=config).getOrCreate()

data = spark.read.parquet("Dados_tratados")

### Amostra dos dados tratados

data.show(25, False)

fechadas = (
    data
    .filter(f.col('flag_ativa') == 0)
    .filter(f.year(f.col("situacao_data")) >= 2010)
)

ativas = data.filter(f.col("flag_ativa") == 1)

data = fechadas.unionByName(ativas)

data.toPandas().info()

data.toPandas().describe()

data.groupBy("flag_ativa").count().show()

data = (
    data
    .withColumn("idade_dias", f.when(f.col("flag_ativa") == 1, f.datediff(f.current_date(), f.col("inicio_data"),))
                              .otherwise(f.datediff(f.col("situacao_data"), f.col('inicio_data'))))
    .withColumn("idade_anos", f.round(f.col("idade_dias") / 365.25, 2))
    
)

data.show(25, False)

data.sort('idade_anos', ascending=False).show(250, False)

df = data.select("capital_social").toPandas()

df.plot.hist(bins=10, logy=True)

data = (
    data
    .withColumn("capital_social_cat", f.when(f.col('capital_social') < 1000, "0_1k")
                                  .when((f.col('capital_social') >= 1000) & (f.col('capital_social') < 10000), "1k_10k")
                                  .when((f.col('capital_social') >= 10000) & (f.col('capital_social') < 100000), "10k_100k")
                                  .otherwise("100k_mais"))
)

data.groupBy("capital_social_cat").count().show()

df = data.groupBy("capital_social_cat").count().sort('count', ascending=False).toPandas()

plt.figure(figsize=(20,10))
sns.catplot(data=df, y='capital_social_cat', x='count', kind='bar', palette="ch:.25", height=5, aspect=2)

df = data.toPandas()

sns.catplot(data=df, kind='count', x='natureza_juridica',  height=5, aspect=2)

fechadas.count()

ativas.count()

fechadas.withColumn("ano_fechamento", f.year(f.col("situacao_data"))).groupBy("ano_fechamento").count().sort("count", ascending=False).show(25, False)

analise = fechadas.withColumn("ano_fechamento", f.year(f.col("situacao_data"))).groupBy("ano_fechamento").count().sort("count", ascending=False).toPandas()

analise[analise['ano_fechamento'] < 2023].sort_values('ano_fechamento').plot(x='ano_fechamento', y='count')

plt.figure(figsize=(20,10))
sns.catplot(data=df, y='capital_social_cat', x='count', kind='bar', palette="ch:.25", height=5, aspect=2)

cnae = spark.read.csv("CNAESECAO_UNIDECODE.csv", header=True, sep=';')

ativas.show(25, False)

fechadas.show(25, False)

data.count()

cnae = cnae.select('cod_cnae', 'grande_area')

data = data.join(cnae, cnae.cod_cnae == data.cnae_principal, how='left').drop('cod_cnae')

data.show(25, False)

df = data.groupBy("grande_area").count().sort('count', ascending=False).toPandas()
sns.catplot(data=df, y='grande_area', x='count', kind='bar', palette="ch:.25")

df = data.select('flag_ativa', 'idade_anos').toPandas()

plt.hist(data=df, x='idade_anos', bins=50)

plt.ylabel("contagem")
plt.xlabel('idade_anos')


df.describe()

df.columns

ceps = spark.read.parquet("ceps_h3.parquet")

ceps.show(25, False)

print(data.count())
data = data.join(ceps, on='cep', how='left').drop_duplicates(['cnpj'])

data.show(25, True)

data.filter(f.col('latitude').isNull()).count()

data.filter(f.col('latitude').isNotNull()).count()

data.columns

data = (
    data
    .withColumn('fechada', f.when(f.col('flag_ativa') == 1, 0).otherwise(1))
)

data.select('cnpj', 'id_matriz', 'fechada', 'idade_anos', 'latitude', 'longitude', 'natureza_juridica', 'capital_social_cat', 'porte', 'grande_area').write.parquet("dados_modelo.parquet", mode='overwrite')

data.select('cnpj', 'h3lv6', 'h3lv7', 'h3lv8', 'h3lv9', 'h3lv10').write.parquet('cnpjs_com_h3.parquet', mode='overwrite')
