==> 4_Modelo_Failure.ipynb <==
from lifelines import WeibullAFTFitter
from pyspark.sql.session import SparkSession, SparkConf
import pyspark.sql.functions as f
import matplotlib.pyplot as plt
plt.rcParams["figure.figsize"] = (10,10)
import pandas as pd
import seaborn as sns
import numpy as np
from lifelines.utils import k_fold_cross_validation

df = pd.read_parquet('dados_modelo.parquet')

df = df.dropna()

cnpj = df['cnpj']

df.columns

df.info()

df = pd.get_dummies(df.drop(columns=['cnpj']),drop_first=True)

df['cnpj'] = cnpj

df

df = df[df['idade_anos'] > 0]

aft = WeibullAFTFitter(penalizer=0.01)

df = df.dropna()

df.columns

df[['latitude', 'longitude']]

df['latitude'] = df['latitude'] + 21.558548
df['longitude'] = df['longitude'] + 45.4394151

df.drop(columns=['grande_area_ELETRICIDADE E GAS'], inplace=True)

aft.fit(df, duration_col='idade_anos', event_col='fechada', ancillary=True)

aft.print_summary()

print(aft.median_survival_time_)
print(aft.mean_survival_time_)

aft.plot()

df['latitude'].min()

aft.plot_partial_effects_on_outcome('latitude', values=np.arange(-2,2,0.5),cmap='coolwarm')

abertas = df[df['fechada'] == 0]

pred = aft.predict_median(abertas)

pred.max()

scores = k_fold_cross_validation(aft, df, duration_col='idade_anos', event_col='fechada', k=3, scoring_method="concordance_index")
print(scores)

ativas = df[df['fechada'] == 0].reset_index()

hazards = aft.predict_cumulative_hazard(ativas,times=[5,10,25,50],conditional_after=ativas['idade_anos']).T.reset_index()

hazards.columns

hazards

ativas['risco_5'] = hazards[5.0]
ativas['risco_10'] = hazards[10.0]
ativas['risco_25'] = hazards[25.0]
ativas['risco_50'] = hazards[50.0]


ativas['tempo_mediano'] = aft.predict_median(ativas, conditional_after=ativas['idade_anos'])

ativas = ativas[ativas['tempo_mediano'] < 100]

h3s = pd.read_parquet('./cnpjs_com_h3.parquet/')

ativas.columns
ativas = ativas[[ 'cnpj', 'risco_5',
       'risco_10', 'risco_25', 'risco_50', 'tempo_mediano', 'idade_anos']]

ativas

ativas.merge(h3s, on='cnpj').to_csv('./AFT_kepler.csv', index=False, mode='w')
