==> 6_Kaplan_Meier.ipynb <==
# Kaplan-Meier
from lifelines import KaplanMeierFitter
from pyspark.sql.session import SparkSession, SparkConf
import pyspark.sql.functions as f
import matplotlib.pyplot as plt
plt.rcParams["figure.figsize"] = (10,5)
import pandas as pd
import seaborn as sns
from lifelines.utils import k_fold_cross_validation

kmf = KaplanMeierFitter()

df = pd.read_parquet('dados_modelo.parquet').dropna()

kmf.fit(durations=df['idade_anos'], event_observed=df['fechada'])

ax = plt.subplot(111)



kmf.plot_survival_function(ax=ax)

plt.title('Total da base')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')

print(kmf.median_survival_time_)

df.columns

nat_jur = list(df['natureza_juridica'].unique())

df['natureza_juridica'].isnull().sum()

ax = plt.subplot(111)

for item in nat_jur:

    nat = (df['natureza_juridica'] == item)
    
    kmf.fit(df[nat]['idade_anos'], df[nat]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)

plt.title('Natureza Jurídica')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')
print(kmf.median_survival_time_)

ax = plt.subplot(111)


portes = list(df['porte'].unique())

for item in portes:
    print(item)

    por = (df['porte'] == item)

    por
    
    kmf.fit(df[por]['idade_anos'], df[por]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)
    print(item, kmf.median_survival_time_)

plt.title('Porte')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')


ax = plt.subplot(111)


areas = list(df['grande_area'].unique())

for item in areas[0:5]:
    print(item)

    cnae = (df['grande_area'] == item)

    cnae
    
    kmf.fit(df[cnae]['idade_anos'], df[cnae]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)
    print(item, kmf.median_survival_time_)

plt.title('CNAE - Grande área')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')

ax = plt.subplot(111)


areas = list(df['grande_area'].unique())

for item in areas[6:11]:
    print(item)

    cnae = (df['grande_area'] == item)

    cnae
    
    kmf.fit(df[cnae]['idade_anos'], df[cnae]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)
    print(item, kmf.median_survival_time_)

plt.title('CNAE - Grande área')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')

ax = plt.subplot(111)


areas = list(df['grande_area'].unique())

for item in areas[12:16]:
    print(item)

    cnae = (df['grande_area'] == item)

    cnae
    
    kmf.fit(df[cnae]['idade_anos'], df[cnae]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)
    print(item, kmf.median_survival_time_)

plt.title('CNAE - Grande área')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')

ax = plt.subplot(111)


areas = list(df['grande_area'].unique())

for item in areas[16:]:
    print(item)

    cnae = (df['grande_area'] == item)

    cnae
    
    kmf.fit(df[cnae]['idade_anos'], df[cnae]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)
    print(item, kmf.median_survival_time_)

plt.title('CNAE - Grande área')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')

ax = plt.subplot(111)


areas = list(df['capital_social_cat'].unique())

for item in areas:
    print(item)

    cnae = (df['capital_social_cat'] == item)

    cnae
    
    kmf.fit(df[cnae]['idade_anos'], df[cnae]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)
    print(item, kmf.median_survival_time_)

plt.title('Capital Social')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')

ax = plt.subplot(111)


areas = list(df['id_matriz'].unique())

for item in areas:
    print(item)

    cnae = (df['id_matriz'] == item)

    cnae
    
    kmf.fit(df[cnae]['idade_anos'], df[cnae]['fechada'], label=item)

    kmf.plot_survival_function(ax=ax)
    print(item, kmf.median_survival_time_)

plt.title('Flag Matriz')
plt.xlabel('Tempo (anos)')
plt.ylabel('Probabilidade de sobrevivência')
